requirements.txt
scikit-learn
numpy
pandas
pyyaml
streamlit
plotly